__NewsLetter_Deploy/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
yfinance==0.2.28
pandas==2.2.3
matplotlib==3.8.2
mplfinance==0.12.10b0
ta==0.10.2
google-generativeai==0.3.2
python-dotenv==1.0.0
email-validator==2.1.0
PyMySQL==1.1.0
gunicorn==21.2.0
psutil==5.9.6
sendgrid==6.11.0
Jinja2==3.1.2
celery==5.3.4
redis==5.0.1 